requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
attrs==25.1.0
beautifulsoup4==4.13.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
eval_type_backport==0.2.2
fastapi==0.110.0
filelock==3.17.0
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
idna==3.10
Jinja2==3.1.3
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
narwhals==1.25.0
numpy==1.26.4
packaging==24.2
pandas==2.2.1
peewee==3.17.8
pillow==10.4.0
platformdirs==4.3.6
plotly==5.19.0
propcache==0.2.1
pyarrow==19.0.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sseclient-py==1.8.0
starlette==0.36.3
tabulate==0.9.0
tenacity==9.0.0
together==1.3.14
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.27.1
webencodings==0.5.1
yarl==1.18.3
yfinance==0.2.37
